+++ src/lib/Layout/PanelSideBarLayout/PanelSideBar/Definitions/PanelItem.ts
@@ -1,6 +1,18 @@
 import { IconProp } from "@fortawesome/fontawesome-svg-core";
 import { ReactNode } from "react";
 
+export type PanelItemOnSideBarCollapseOptions = {
+  /**
+   * Whether the sidebar maintains the panel item icon visible on collapsing.
+   */
+  showIcon?: boolean;
+
+  /**
+   * The icon to be displayed when the active panel item has `showIcon` enabled, the sidebar is collapsed and the panel item does not have any icon.
+   */
+  fallbackIcon?: IconProp;
+} 
+
 export type PanelItem<TPanelItemId extends string, TPanelItem = Record<string, unknown>> = TPanelItem & {
   /**
    * The panel icon.
@@ -41,12 +53,12 @@ export type PanelItem<TPanelItemId extends string, TPanelItem = Record<string, u
    */
   active?: boolean;
   /**
-   * Whether collapse only with icon.
-   */
+  * Whether collapse only with icon.
+  */
   collapseIconOnly?: boolean;
 
   /**
-   * Whether the sidebar maintains the panel items icons visible on collapsing.
+   * The panel item options once the sidebar gets collapsed.
    */
-  showIconsOnCollapse?: boolean;
+  onSidebarCollapseOptions?: PanelItemOnSideBarCollapseOptions;
 };