+++ src/lib/Layout/PanelSideBarLayout/PanelSideBar/PanelSideBarItem.tsx
@@ -1,11 +1,10 @@
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import classNames from "classnames";
-import { useRef, useEffect, ReactNode } from "react";
+import { useRef, useEffect } from "react";
 import { Collapse, NavItem } from "reactstrap";
 import { PanelItem } from "./../PanelSideBar/Definitions/PanelItem";
 import { usePanelSideBarContext } from "./Context/PanelSideBarContext";
 import { hasActiveChildren } from "./Utils/getActivePanel";
-import { IconProp } from "@fortawesome/fontawesome-svg-core";
 
 export interface PanelSideBarItemProps<TPanelItemId extends string, TPanelItem> {
   children: PanelItem<TPanelItemId, TPanelItem>;
@@ -14,25 +13,36 @@ export interface PanelSideBarItemProps<TPanelItemId extends string, TPanelItem> 
   isParentHidden?: boolean;
 }
 
-const PanelSidebarItemNavLink = ({ icon, title, collapsedWithIcons }: { icon?: IconProp, title: ReactNode, collapsedWithIcons?: boolean }) => {
-  const iconClassName = collapsedWithIcons ? "ms-1 me-3 p-1" : "me-2";
+const PanelSidebarItemNavLink = <TPanelItemId extends string, TPanelItem>({
+  item,
+  collapsedWithIcon,
+  className,
+}: {
+  item:  PanelItem<TPanelItemId, TPanelItem>;
+  collapsedWithIcon?: boolean;
+  className?: string;
+}) => {
+  const { icon, title, onSidebarCollapseOptions } = item;
+  const panelIconClassName = collapsedWithIcon ? "ms-1 me-3 p-1" : "me-2";
+  const displayIcon = icon || (collapsedWithIcon && onSidebarCollapseOptions?.fallbackIcon);
+
   return (
-    <span className="nav-link">
-      {icon && <FontAwesomeIcon icon={icon} className={iconClassName} />}
-      {collapsedWithIcons && icon ? "" : title}
+    <span className={className}>
+      {displayIcon && <FontAwesomeIcon icon={displayIcon} className={panelIconClassName} />}
+      {!collapsedWithIcon || !displayIcon ? title : ""}
     </span>
   );
-}
+};
 
 // eslint-disable-next-line complexity
 const PanelSideBarItem = <TPanelItemId extends string, TPanelItem>(props: PanelSideBarItemProps<TPanelItemId, TPanelItem>) => {
   const { depth = 0, children: item, isParentHidden = false } = props;
-  const { LinkRenderer, toggledMenuItemIds, toggleMenuItem, hiddenMenuItemIds, activePanelShowIconsOnCollapse, isSidebarOpen } = usePanelSideBarContext<TPanelItemId, TPanelItem>();
+  const { LinkRenderer, toggledMenuItemIds, toggleMenuItem, hiddenMenuItemIds, isIconShownOnSidebarCollapse, isSidebarOpen } = usePanelSideBarContext<TPanelItemId, TPanelItem>();
   const hasItems = !!item.children?.length;
   const isActive = (hasItems && item.children && hasActiveChildren(item.children)) || item.active;
   const isOpen = toggledMenuItemIds?.includes(item.id);
   const scrollToActiveItemRef = useRef<HTMLDivElement>(null);
-  const collapsedWithIcons = activePanelShowIconsOnCollapse && !isSidebarOpen;
+  const collapsedWithIcon = isIconShownOnSidebarCollapse && !isSidebarOpen;
 
   useEffect(() => {
     if (scrollToActiveItemRef.current && isActive) {
@@ -50,20 +60,20 @@ const PanelSideBarItem = <TPanelItemId extends string, TPanelItem>(props: PanelS
           }
         }}
         className={classNames({ "menu-open": isOpen, active: isActive })}
-        style={{ paddingLeft: !collapsedWithIcons && depth ? `${depth + 1}rem` : undefined }}
+        style={{ paddingLeft: !collapsedWithIcon && depth ? `${depth + 1}rem` : undefined }}
       >
         <div ref={scrollToActiveItemRef}>
           {hasItems ? (
             <div className={classNames("d-flex flex-row", { "justify-content-between": item.collapseIconOnly })}>
               {item.collapseIconOnly && (
                 <LinkRenderer item={item}>
-                  <PanelSidebarItemNavLink icon={item.icon} title={item.title} collapsedWithIcons={collapsedWithIcons} />
+                  <PanelSidebarItemNavLink<TPanelItemId, TPanelItem> className="nav-link" item={item} collapsedWithIcon={collapsedWithIcon} />
                 </LinkRenderer>
               )}
 
               <a
                 role="button"
-                className={classNames("nav-link", { "w-100": !item.collapseIconOnly }, { "dropdown-toggle": hasItems && !collapsedWithIcons })}
+                className={classNames("nav-link", { "w-100": !item.collapseIconOnly }, { "dropdown-toggle": hasItems && !collapsedWithIcon })}
                 onClick={() => {
                   if (item.collapseIconOnly) {
                     toggleMenuItem(item.id);
@@ -71,14 +81,14 @@ const PanelSideBarItem = <TPanelItemId extends string, TPanelItem>(props: PanelS
                 }}
               >
                 {!item.collapseIconOnly && (
-                  <PanelSidebarItemNavLink icon={item.icon} title={item.title} collapsedWithIcons={collapsedWithIcons} />
+                  <PanelSidebarItemNavLink<TPanelItemId, TPanelItem> item={item} collapsedWithIcon={collapsedWithIcon} />
                 )}
               </a>
             </div>
           ) : (
             <>
               <LinkRenderer item={item}>
-                <PanelSidebarItemNavLink icon={item.icon} title={item.title} collapsedWithIcons={collapsedWithIcons} />
+              <PanelSidebarItemNavLink<TPanelItemId, TPanelItem> className="nav-link" item={item} collapsedWithIcon={collapsedWithIcon} />
               </LinkRenderer>
             </>
           )}